Get the server up and running
