+++ README.md
@@ -1,2 +1,4 @@
 # cursed-objects
 Museum objects. But cursed... Mwahahahaha!
+
+Run: `node server.js`
