+++ views/index.ejs
@@ -1,17 +1,27 @@
 <!DOCTYPE html>
 <html>
-    <head>
-        <meta charset="utf-8">
-        <title>Cursed Objects</title>
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" type="text/css" media="screen" href="/css/style.css" />
-        <script src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
-        <script src="../bundle.js"></script>
-        <script src="main.js"></script>
-    </head>
-    <body>
-        <div id="container">
-        </div>
+<head>
+  <meta charset="utf-8">
+  <title>Cursed Objects</title>
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="stylesheet" type="text/css" media="screen" href="/css/style.css" />
+  <script src="https://code.jquery.com/jquery-3.3.1.min.js"></script>
+  <!-- <script src="../bundle.js"></script> -->
+  <!-- <script src="main.js"></script> -->
+</head>
+<body>
+  <div id="container">
+    <form action="/" method="post">
+      <input name="artefactIdentifier" type="text" class="" placeholder="Enter an ID" required>
+      <input type="submit" class="" value="Go!">
+    </form>
+    <% if (story !== null){ %>
+      <p><%= story %></p>
+      <% } %>
+      <% if (error !== null){ %>
+        <p><%= error %></p>
+        <% } %>
+      </div>
     </body>
-</html>
+    </html>
